requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2022.6
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0.post1
xgboost==1.7.1
